+++ library/__init__.py
@@ -12,5 +12,5 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-__author__ = 'Luigi Mori (@jtschichold), Ivan Bojer (@ivanbojer)'
+__author__ = 'Luigi Mori (@jtschichold), Ivan Bojer (@ivanbojer), Vinay Venkataraghavan (@vinayvenkat)'
 __version__ = '1.0.3'