+++ HISTORY.rst
@@ -4,6 +4,15 @@
 History
 =======
 
+V1.0.5
+------
+
+- panos_sag: Added the ability to add / delete static address groups.
+- panos_dag: Converted the module to use pandevice
+             Also added the ability to perform create / delete / list
+- panos_dag_tags: A new module to create registered IP to tag associations
+                  Implemented the ability to create / delete / list IP to tag associations
+
 V1.0.4
 ------